# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Weather API",
    "description": "JSON REST API weather data readings"
  },
  "servers": [
    {
      "url": "http://localhost:8088"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Users - GET"
        ],
        "summary": "Get a list with all users",
        "description": "",
        "parameters": [
          {
            "name": "authenticationKey",
            "in": "header",
            "required": true,
            "example": "be39783e-0aaa-4fc8-b0d8-5e3f0a8b1465",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Users - POST"
        ],
        "summary": "Create a user",
        "description": "",
        "parameters": [
          {
            "name": "authenticationKey",
            "in": "header",
            "required": true,
            "example": "be39783e-0aaa-4fc8-b0d8-5e3f0a8b1465",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "User Login",
          "content": {
            "application/json": {
              "schema": {
                "email": "string",
                "password": "number",
                "role": "string",
                "firstName": "string",
                "lastName": "string"
              },
              "example": {
                "email": "email@server.com",
                "password": "password",
                "role": "role",
                "firstName": "Firstname",
                "lastName": "Lastname"
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users - GET"
        ],
        "summary": "Get a specific user by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authenticationKey",
            "in": "header",
            "required": true,
            "example": "be39783e-0aaa-4fc8-b0d8-5e3f0a8b1465",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/by-key/{authenticationKey}": {
      "get": {
        "tags": [
          "Users - GET"
        ],
        "summary": "Get a user by the authentication key ",
        "description": "",
        "parameters": [
          {
            "name": "authenticationKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authenticationKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "be39783e-0aaa-4fc8-b0d8-5e3f0a8b1465"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Users - POST"
        ],
        "summary": "User Login",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "User Login",
          "content": {
            "application/json": {
              "schema": {
                "email": "string",
                "password": "number"
              },
              "example": {
                "email": "email@server.com",
                "password": "password"
              }
            }
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "Users - POST"
        ],
        "summary": "User Logout",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "User Logout",
          "content": {
            "application/json": {
              "schema": {
                "authenticationKey": "string"
              },
              "example": {
                "authenticationKey": "22ebfdea-7535-43b4-9436-7ca2ab9f2408"
              }
            }
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "tags": [
          "Users - POST"
        ],
        "summary": "Register a user",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Register a new User",
          "content": {
            "application/json": {
              "schema": {
                "email": "string",
                "password": "number",
                "firstName": "string",
                "lastName": "string"
              },
              "example": {
                "email": "email@server.com",
                "password": "password",
                "firstName": "Firstname",
                "lastName": "Lastname"
              }
            }
          }
        }
      }
    },
    "/users/update": {
      "patch": {
        "tags": [
          "Users - PATCH"
        ],
        "summary": "Only Update a value for a specific a user",
        "description": "",
        "parameters": [
          {
            "name": "authenticationKey",
            "in": "header",
            "required": true,
            "example": "be39783e-0aaa-4fc8-b0d8-5e3f0a8b1465",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Update a existing User ",
          "content": {
            "application/json": {
              "schema": {
                "id": "string",
                "email": "string",
                "password": "number",
                "role": "string",
                "firstName": "string",
                "lastName": "string",
                "authenticationKey": "string"
              },
              "example": {
                "id": "641a6bf00e2c74fca47ea4a1",
                "email": "email@server.com",
                "password": "password",
                "role": "role",
                "firstName": "Firstname",
                "lastName": "Lastname",
                "authenticationKey": "authenticationKey"
              }
            }
          }
        }
      }
    },
    "/user/role": {
      "put": {
        "tags": [
          "Users - PUT"
        ],
        "summary": "Update user role by date.",
        "description": "",
        "parameters": [
          {
            "name": "authenticationKey",
            "in": "header",
            "required": true,
            "example": "be39783e-0aaa-4fc8-b0d8-5e3f0a8b1465",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Update access level for at least two users in the same query, based on a date range in which the users were created.  Start Date in the format of: YYYY-MM-DD,   End Date in the format of: YYYY-MM-DD and   New User Role. Admin, Teacher or Student ",
          "content": {
            "application/json": {
              "schema": {
                "startDate": "string",
                "endDate": "string",
                "newRole": "string"
              },
              "example": {
                "startDate": "2023-05-01",
                "endDate": "2023-05-07",
                "newRole": "Teacher"
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "put": {
        "tags": [
          "Users - PUT"
        ],
        "summary": "Create or Override a value for a specific user by ID.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authenticationKey",
            "in": "header",
            "required": true,
            "example": "be39783e-0aaa-4fc8-b0d8-5e3f0a8b1465",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Create or Override a value for a specific user by ID",
          "content": {
            "application/json": {
              "schema": {
                "email": "string",
                "password": "number",
                "firstName": "string",
                "lastName": "string"
              },
              "example": {
                "email": "email@server.com",
                "password": "password",
                "firstName": "Firstname",
                "lastName": "Lastname"
              }
            }
          }
        }
      }
    },
...
  }
}